refactor(phase4): move capital dead letter payload to core
